Move comparators to own module
